userModel moved to repository/storage/postgres/orm/sequelize/models
